+++ examples/calculator/tests.py
@@ -28,7 +28,7 @@ class TestRpcThread(test.TestCase):
         return cls.worker.arbiter.send(cls.worker,'kill_actor',cls.app.mid)
         
     def setUp(self):
-        self.p = rpc.JsonProxy(self.uri)
+        self.p = rpc.JsonProxy(self.uri, timeout = 10)
         
     def testHandler(self):
         s = self.app
@@ -78,6 +78,13 @@ class TestRpcThread(test.TestCase):
         
     def testInternalError(self):
         self.assertRaises(rpc.InternalError,self.p.calc.divide,'ciao','bo')
+        
+    def testpaths(self):
+        '''Ftech a sizable ammount of data'''
+        result = self.p.calc.randompaths(num_paths = 200,
+                                         size = 1000,
+                                         mu = 1, sigma = 2)
+        self.assertTrue(result)
         
 
 class TestRpcProcess(TestRpcThread):