+++ pulsar/lib/fallback/http.py
@@ -1,3 +1,30 @@
+'''Original code from https://github.com/benoitc/http-parser
+
+2011 (c) Benoit Chesneau <benoitc@e-engura.org>
+
+Permission is hereby granted, free of charge, to any person
+obtaining a copy of this software and associated documentation
+files (the "Software"), to deal in the Software without
+restriction, including without limitation the rights to use,
+copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the
+Software is furnished to do so, subject to the following
+conditions:
+
+The above copyright notice and this permission notice shall be
+included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+OTHER DEALINGS IN THE SOFTWARE.
+
+Modified and adapted for pulsar.
+'''
 import os
 import re
 import sys
@@ -53,7 +80,7 @@ class HttpParser(object):
         self._buf = [] 
         self._version = None
         self._method = None
-        self._protocol = None
+        self._server_protocol = None
         self._status_code = None
         self._status = None
         self._reason = None
@@ -102,7 +129,7 @@ class HttpParser(object):
         return self._headers
     
     def get_protocol(self):
-        return self._protocol
+        return self._server_protocol
     
     def get_body(self):
         return self._body