+++ pulsar/net/client/__init__.py
@@ -7,6 +7,7 @@ from .std import HttpClientHandler, HttpClient1, urlencode,\
 __all__ = ['HttpClientHandler',
            'HttpClient',
            'HttpClientResponse',
+           'setDefaultClient',
            'urlencode',
            'responses']
 
@@ -20,6 +21,13 @@ except ImportError:
 
 
 form_headers = {'Content-type': 'application/x-www-form-urlencoded'}
+
+_DefaultClient = (2,1)
+
+
+def setDefaultClient(v):
+    global _DefaultClient
+    _DefaultClient = v
 
 
 class AsyncHttpClient(object):
@@ -35,7 +43,7 @@ class AsyncHttpClient(object):
         
 
 def HttpClient(cache = None, proxy_info = None,
-               timeout = None, type = 1, ioloop = None,
+               timeout = None, type = None, ioloop = None,
                async = False, handle_cookie = False):
     '''Factory of :class:`HttpClientHandler` instances.
 It can build a synchronous or an asynchronous handler build on top
@@ -47,9 +55,17 @@ of the :class:`pulsar.IOLoop`.
 :parameter type: Request handler implementation. Default ``1``.
 :parameter async: Synchronous or Asynchronous. Default ``False``.
 '''
-    if type not in HttpClients:
-        raise ValueError('HttpClient{0} not available'.format(type))
-    client = HttpClients[type]
+    type = type or _DefaultClient
+    if isinstance(type,int):
+        type = (type,)
+    ctype = None
+    for t in type:
+        if t in HttpClients:
+            ctype = t
+            break
+    if not ctype:
+        raise ValueError('HttpClients {0} not available'.format(type))
+    client = HttpClients[ctype]
     proxy = proxy_info
     if proxy is None:
         proxy = getproxies_environment()