+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Development
 =======================================
-* **61 regression tests**
+* Fixed chunked transfer encoding.
+* Fixed critical bug in socket server :class:`pulsar.Mailbox`. Each client connections
+  has its own buffer.
+* **64 regression tests**
 
 Version 0.2.0 - 2011-Nov-05
 =======================================